requirements.txt
selenium
webdriver-manager
python-dotenv
requests
litellm
# Add other dependencies here as needed
